backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.25
psycopg[binary]==3.1.12
alembic==1.13.1
python-dotenv==1.0.1
black==24.4.2
ruff==0.4.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
pydantic[email]==2.7.1
